fix SilenceProcessingTest when number of outputs > number of inputs
this loop sets a silence flag for all of the channels in
processData.outputs[busIndex]. however, it takes the channel count from
processData.inputs[busIndex].numChannels.

when there are more output buses than input buses in a plugin, this
resulted in an invalid memory read past the end of processData.inputs.